feat(f16): stage 3 -- fp16 storage for Monarch's zs activation cache
Converts FwdCache.zs from Vec<f32> to Vec<half::f16> -- the single
largest buffer BufPool manages, written once per token per block in
forward and read once per token per block in backward's phase1. Unlike
the tied embedding table (stage 2), zs has no gradient/accumulator
counterpart needing a persistent fp32 "master copy", so this is a cleaner
case for both bandwidth *and* memory: no footgun like apply_grad's
transient buffer from stage 2.

Design: the hot apply_block/apply_block_avx2 kernel (called once per
token per block -- hundreds of thousands of times per step at production
scale) is left completely untouched, fp32 throughout. Its callers (the
three forward_batch/forward_rows_batch/forward_cols_batch methods'
per-token closures) write into a small reused f32 scratch buffer as
before, then convert-and-store into the fp16 zs cache right after
apply_block returns -- conversion happens once at the caller boundary,
never inside the dense AVX2 kernel itself. Symmetrically,
backward_block_phase1 (also untouched) reads from a small reused f32
scratch that the three backward_*_batch methods' run_range closures
populate via one conversion call per block, right before calling it.

BufPool gained a parallel fp16 free-list (take_f16_uninit/give_f16) since
an f32 and half::f16 buffer of the same element count aren't
interchangeable.

The non-batched single-token forward/backward/forward_rows/forward_cols/
backward_rows/backward_cols methods (SharedMonarchMatmul and
SharedMonarchProj) are only used by tests and research bins (gate_c.rs),
never the production hot path -- for those, used a simpler bulk-convert-
at-the-function-boundary pattern instead of restructuring their loops.

Correctness: 106/106 tests pass, including the critical gradcheck suite
(backward_batch_matches_summed_looped_backward,
shared_monarch_proj_gradcheck, phase1_plus_contract_matches_backward_block_hoisted,
and 11 others) -- all still validated against untouched fp32 reference
implementations, now just reading/writing fp16-stored zs at the boundary.

Measured, FULL=1 production shape (96 layers, seq=512), interleaved A/B
against the pre-fp16 commit (two pairs, consistent both times): total
step time ~34.4-35.0s/step -> ~31.5-31.7s/step, a combined ~8-9%
reduction across both fp16 stages. Peak RSS: 2.05GB (stage 2 alone,
regressed by the apply_grad footgun) -> 1.94GB (stage 3 landed), close to
the original pre-fp16 baseline of 1.99GB -- zs's own memory reduction is
now showing through, unlike stage 2's embed table which had no net
peak-memory win.

+++ src/model/attn_proj.rs
@@ -176,7 +176,7 @@ impl AttnProj {
 
     /// Slice a single token's cache slice out of a batched `forward_batch`
     /// cache (no-op / empty for the dense path).
-    pub fn zs_at<'a>(&self, cache: &'a FwdCache, token: usize) -> &'a [f32] {
+    pub fn zs_at<'a>(&self, cache: &'a FwdCache, token: usize) -> &'a [half::f16] {
         match &self.kind {
             ProjKind::Monarch { proj } => proj.zs_at(cache, token),
             ProjKind::Dense { .. } => &[],
@@ -185,7 +185,7 @@ impl AttnProj {
 
     /// VJP. `dy` is the gradient w.r.t. the projection output; `zs` is this
     /// token's cache slice from `forward_batch` (ignored by the dense path).
-    pub fn backward(&self, d1: &[f32], d2: &[f32], x: &[f32], zs: &[f32], dy: &[f32]) -> ProjGrads {
+    pub fn backward(&self, d1: &[f32], d2: &[f32], x: &[f32], zs: &[half::f16], dy: &[f32]) -> ProjGrads {
         debug_assert_eq!(x.len(), self.in_, "projection input shape mismatch");
         debug_assert_eq!(dy.len(), self.out, "projection grad-output shape mismatch");
         match &self.kind {
